split event out from log
