+++ crates/analyzer/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod error;
+pub mod event;
 pub mod log;
 pub mod rules;
 pub mod users;
